#12795: Move to ttnn/cpp/ttnn/operations/numpy
