+++ tests/tt_eager/ops/test_bcast_op.cpp
@@ -7,7 +7,7 @@
 #include "ttnn/operations/data_movement/bcast/bcast.hpp"
 #include "common/constants.hpp"
 #include "third_party/magic_enum/magic_enum.hpp"
-#include <ttnn/operations/functions.hpp>
+#include <ttnn/operations/numpy/functions.hpp>
 
 #include <algorithm>
 #include <functional>
