+++ ttnn/cpp/ttnn/operations/data_movement/reshape/reshape.cpp
@@ -7,7 +7,7 @@
 #include "ttnn/run_operation.hpp"
 #include "reshape.hpp"
 #include "tt_metal/common/constants.hpp"
-#include <ttnn/operations/functions.hpp>
+#include <ttnn/operations/numpy/functions.hpp>
 #include "ttnn/operations/experimental/auto_format/auto_format.hpp"
 #include "ttnn/tensor/tensor_utils.hpp"
 #include "device/reshape_op.hpp"
